Reduce casts in deserialization

+++ serialization.go
@@ -50,10 +50,10 @@ func SerializeChunk(c Chunk) []byte {
 }
 
 func DeserializeChunk(b []byte) *Chunk {
-	x := int16(binary.LittleEndian.Uint16(b[0:2]))
-	y := int16(binary.LittleEndian.Uint16(b[2:4]))
+	x := int(int16(binary.LittleEndian.Uint16(b[0:2])))
+	y := int(int16(binary.LittleEndian.Uint16(b[2:4])))
 
-	c := NewChunk(pixel.R(float64(x)*float64(ChunkSize), float64(y)*float64(ChunkSize), float64(x+1)*float64(ChunkSize), float64(y+1)*float64(ChunkSize)))
+	c := NewChunk(pixel.R(float64(x*ChunkSize), float64(y*ChunkSize), float64((x+1)*ChunkSize), float64((y+1)*ChunkSize)))
 
 	byteIndex := 4
 	for x := range c.Tiles {
